+++ backend/src/main/java/com/team3/airdnd/accommodation/controller/AccommodationController.java
@@ -1,0 +1,24 @@
+package com.team3.airdnd.accommodation.controller;
+
+import com.team3.airdnd.accommodation.dto.AccommodationResponseDto;
+import com.team3.airdnd.accommodation.service.AccommodationService;
+import com.team3.airdnd.global.dto.ResponseDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/accommodations")
+public class AccommodationController {
+
+    private final AccommodationService accommodationService;
+
+    @GetMapping("/{accommodation-id}")
+    public ResponseDto<AccommodationResponseDto.AccommodationDetailDto> getAccommodationDetail(@PathVariable("accommodation-id") Long id) {
+        AccommodationResponseDto.AccommodationDetailDto detailDto = accommodationService.getAccommodationDetail(id);
+        return ResponseDto.ok(detailDto);
+    }
+}
