+++ backend/src/main/java/com/team3/airdnd/accommodation/dto/AccommodationRequestDto.java
@@ -1,0 +1,4 @@
+package com.team3.airdnd.accommodation.dto;
+
+public class AccommodationRequestDto {
+}
