+++ backend/src/main/java/com/team3/airdnd/accommodation/dto/AccommodationResponseDto.java
@@ -1,0 +1,4 @@
+package com.team3.airdnd.accommodation.dto;
+
+public class AccommodationResponseDto {
+}
