+++ backend/src/main/java/com/team3/airdnd/accommodation/service/AccommodationService.java
@@ -1,9 +1,9 @@
-package com.team3.airdnd.accomodation.service;
+package com.team3.airdnd.accommodation.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class AccomodationService {
+public class AccommodationService {
 }
